Pipe Character and Video Title
Reddit apparently lacks an escape for pipes ("|") in tables. Videos with pipes in the title would cause weird weird formatting. The closest character is the vertical line for line drawing, so it now replaces "|" with "│". It's not pretty but it works.

Also added the YouTube video title to the print_video_info() output.